+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Bug Fixes
 
 - Made several improvements to macro hygiene by renaming local variable bindings during expansion. Added a unit test module covering many test cases.
 - Fixed many functions including `utf8->string`, `string->utf8`, `bytevector-copy`, `list->vector`, and `list->string` to heap allocate objects that exceed the maximum size for objects on the stack.
+- Prevent a compiler error when there is only one argument to `+` or `*`.
 
 ## 0.7.1 - December 21, 2017